+++ install_zabbix.sh
@@ -8,15 +8,18 @@ export old_mism_version=$MISM_BULLSEQUANA_EDGE_VERSION
 
 if [ ! -z $MISM_BULLSEQUANA_EDGE_VERSION ]
 then
-  if [ $old_mism_version != $MISM_BULLSEQUANA_EDGE_VERSION  ]
+  if [ ! -z $old_mism_version ]
   then
-    rm -f bullsequana-edge-system-management_zabbix-web.$old_mism_version.tar
-    rm -f bullsequana-edge-system-management_zabbix-web.$old_mism_version.tar
-    rm -f bullsequana-edge-system-management_zabbix-agent.$old_mism_version.tar
-    rm -f zabbix-server-pgsql.$old_mism_version.tar
-    rm -f zabbix-web-nginx-pgsql.$old_mism_version.tar
-    rm -f zabbix-agent.$old_mism_version.tar
-    rm -f postgres.$old_mism_version.tar
+    if [ $old_mism_version != $MISM_BULLSEQUANA_EDGE_VERSION  ]
+    then
+      rm -f bullsequana-edge-system-management_zabbix-web.$old_mism_version.tar
+      rm -f bullsequana-edge-system-management_zabbix-web.$old_mism_version.tar
+      rm -f bullsequana-edge-system-management_zabbix-agent.$old_mism_version.tar 
+      rm -f zabbix-server-pgsql.$old_mism_version.tar
+      rm -f zabbix-web-nginx-pgsql.$old_mism_version.tar
+      rm -f zabbix-agent.$old_mism_version.tar
+      rm -f postgres.$old_mism_version.tar
+    fi
   fi
 fi
